[SPARK-41466][BUILD] Change Scala Style configuration to catch AnyFunSuite instead of FunSuite
### What changes were proposed in this pull request?

This PR is technically a followup of https://github.com/apache/spark/pull/29196 that changes Scala Style configuration to catch `AnyFunSuite` instead of `FunSuite`.

### Why are the changes needed?

To keep the original intention of the Scala style configuration.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

Manually tested with `./dev/lint-scala`. CI in this PR should test it out too.

Closes #39004 from HyukjinKwon/SPARK-41466.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>