[SPARK-41187][CORE] LiveExecutor MemoryLeak in AppStatusListener when ExecutorLost happen
### What changes were proposed in this pull request?
Ignore the SparkListenerTaskEnd with Reason "Resubmitted" in AppStatusListener to avoid memory leak

### Why are the changes needed?
For a long running spark thriftserver,  LiveExecutor will be accumulated in the deadExecutors HashMap and cause message event queue processing slowly.
For a every task, actually always sent out a `SparkListenerTaskStart` event and a `SparkListenerTaskEnd` event, they are always pairs. But in a executor lost situation, it send out event like following steps.

a) There was a pair of task start and task end event which were fired for the task (let us call it Tr)
b) When executor which ran Tr was lost, while stage is still running, a task end event with reason `Resubmitted`  is fired for Tr.
c) Subsequently, a new task start and task end will be fired for the retry of Tr.

The processing of the `Resubmitted` task end event in AppStatusListener can lead to negative `LiveStage.activeTasks` since there's no corresponding `SparkListenerTaskStart` event for each of them. The negative activeTasks will make the stage always remains in the live stage list as it can never meet the condition activeTasks == 0. This in turn causes the dead executor to never be cleaned up if that live stage's submissionTime is less than the dead executor's removeTime( see isExecutorActiveForLiveStages). Since this kind of `SparkListenerTaskEnd` is useless here, we simply ignore it.

Check  [SPARK-41187](https://issues.apache.org/jira/browse/SPARK-41187) for evidences.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
New UT Added
Test in thriftserver env

### The way to reproduce
I try to reproduce it in spark shell, but it is a little bit handy
1.  start spark-shell , set spark.dynamicAllocation.maxExecutors=2 for convient
` bin/spark-shell  --driver-java-options "-agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=8006"`
2. run a job with shuffle
 `sc.parallelize(1 to 1000, 10).map { x => Thread.sleep(1000) ; (x % 3, x) }.reduceByKey((a, b) => a + b).collect()`
3. After some ShuffleMapTask finished, kill one or two executor to let tasks resubmitted
4. check by heap dump or debug or log

Closes #38702 from wineternity/SPARK-41187.

Authored-by: yuanyimeng <yuanyimeng@youzan.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>