[SPARK-41520][SQL] Split AND_OR TreePattern to separate AND and OR TreePatterns
### What changes were proposed in this pull request?

Removed `AND_OR` TreePattern, replaced with separate `AND` TreePattern and `OR` TreePattern.

### Why are the changes needed?

This way we can be more fine-grained with how we match for AND/OR patterns, since in many cases we want to treat them separately.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

No tests were added as there are not any existing unit tests for TreePatterns. Instead, I just made sure the existing relevant tests (e.g. `BooleanSimplificationSuite`) passed.

Closes #39064 from kelvinjian-db/SPARK-41520-and-or-treepattern.

Authored-by: Kelvin Jiang <kelvin.jiang@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>