[SPARK-41387][SS] Assert current end offset from Kafka data source for Trigger.AvailableNow
### What changes were proposed in this pull request?

This PR proposes to assert the end offset of current batch for Kafka data source with Trigger.AvailableNow, via fetching the latest offset from known topic partitions and compare with end offset. Kafka data source will throw error when the latest information from topic-partition is a loss e.g. topic-partition is missing, latest offset is less than the offset in end offset.

This way we construct a safeguard for Trigger.AvailableNow to avoid hanging, although it won't prevent 100% of cases for future bugs.

### Why are the changes needed?

This change will let Kafka data source do its best effort to fail fast when the state of the query is not able to reach the end state of Trigger.AvailableNow.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

New test cases.

Closes #38911 from HeartSaVioR/SPARK-41387.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>