[SPARK-41498] Propagate metadata through Union
### What changes were proposed in this pull request?

Propagate the `metadataOutput` through the `Union` operator. Propagate only when the `metadataOutput` in all children matches structurally by data type and name. This is done as the metadata is transparent to the user, and we want to ensure that it is only propagated when the metadata output of the children is exactly the same, to avoid unexpected behavior.

### Why are the changes needed?

Currently, the Union operator does not propagate any metadata output. This makes it impossible to access any metadata if a Union operator is used, even though the children have the exact same metadata output.
Example:

```
val df1 = spark.read.load(path1)
val df2 = spark.read.load(path2)
df1.union(df2).select("_metadata.file_path"). // <-- fails
```
The issue here is if the datasets unioned do not contain the same metadata, we won't be able to union them. Hence, we will only propagate the metadata if all children have the same metadata (by data type and name), and otherwise do not propagate it. The code for this is similar to the code used in `CheckAnalysis` to check if the output of a Union can be correctly computed.

### Does this PR introduce _any_ user-facing change?

Yes, users are now able to select metadata columns through unions.

### How was this patch tested?

Unit tests to check that metadata columns can now be accessed through `Union` if the metadata output is the same, and that it is not propagated if the metadata output does not match.

Closes #38941 from fred-db/union-metadata-propagation.

Lead-authored-by: fred-db <fredrik.klauss@databricks.com>
Co-authored-by: Fredrik Klauss <fredrik.klauss@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>