[SPARK-41533][CONNECT] Proper Error Handling for Spark Connect Server / Client
### What changes were proposed in this pull request?
This PR improves the error handling on the Spark Connect server and client side. First, this patch moves the error handling logic on the server into a common error handler partial function that differentiates between the internal Spark errors and other runtime errors.

For custom Spark exceptions, the actual internal error is wrapped into a Google RPC Status and sent as trailing metadata to the client.

On the client side, similarly, the error handling is moved into a common function. All GRPC errors are wrapped into custom exceptions to avoid presenting the user with confusing GRPC errors. If available the attached RPC status is extracted and added to the exception.

Lastly, this patch adds basic logging functionality that can be enabled using the environment variable `SPARK_CONNECT_LOG_LEVEL` and can be set to `info`, `warn`, `error`, and `debug`.

### Why are the changes needed?
Usability

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT

Closes #39212 from grundprinzip/SPARK-41533.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>