[SPARK-41685][UI] Support Protobuf serializer for the KVStore in History server
### What changes were proposed in this pull request?

Support configurable(`spark.history.store.serializer`) serializer for the KVStore in History server. The value can be `JSON` or `PROTOBUF`.

### Why are the changes needed?

Support the fast and compact protobuf serializer in Spark history server(SHS), so that:
* Improve the performance of KVStore IO
* Make it possible for SHS to read the live UI rocksdb instance.

### Does this PR introduce _any_ user-facing change?

Yes, introduce a new configuration `spark.history.store.serializer` for setting the serializer for the KVStore in History server, which can be `JSON` or `PROTOBUF`

### How was this patch tested?

UT

Closes #39202 from gengliangwang/shsProtobuf.

Authored-by: Gengliang Wang <gengliang@apache.org>
Signed-off-by: Gengliang Wang <gengliang@apache.org>