[SPARK-41810][CONNECT] Infer names from a list of dictionaries in SparkSession.createDataFrame
### What changes were proposed in this pull request?

This PR proposes to support to infer field names when the input data is the list of dictionaries in `SparkSession.createDataFrame`.

For example,

```python
spark.createDataFrame([{"course": "dotNET", "earnings": 10000, "year": 2012}]).show()
```

**Before**:

```
+------+-----+----+
|    _1|   _2|  _3|
+------+-----+----+
|dotNET|10000|2012|
+------+-----+----+
```

**After**:

```
+------+--------+----+
|course|earnings|year|
+------+--------+----+
|dotNET|   10000|2012|
+------+--------+----+
```

### Why are the changes needed?

To match the behaviour with the regular PySpark.

### Does this PR introduce _any_ user-facing change?

No to end users because Spark Connect has not been released.

### How was this patch tested?

Unittest was added.

Closes #39344 from HyukjinKwon/SPARK-41746.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>