[SPARK-41311][SQL][TESTS] Rewrite test RENAME_SRC_PATH_NOT_FOUND to trigger the error from user space
### What changes were proposed in this pull request?

This updates the test for error class RENAME_SRC_PATH_NOT_FOUND in QueryExecutionErrorsSuite to trigger the error with user-facing APIs.

### Why are the changes needed?

To test the code more directly.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

The only change is the modification of the RENAME_SRC_PATH_NOT_FOUND test.

### License

This contribution is my original work. I license the work to the project under the project’s open source license.

Closes #39348 from ibuder/SPARK-41311.

Authored-by: Immanuel Buder <immanuel_buder@intuit.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>