[SPARK-41859][SQL] CreateHiveTableAsSelectCommand should set the overwrite flag correctly
### What changes were proposed in this pull request?

`CreateHiveTableAsSelectCommand` is used in two APIs:
1. SQL CTAS
2. `DataFrameWriter.saveAsTable`

For SQL CTAS, if the table already exists, we either do nothing if `IF NOT EXISTS` is specified, or fail otherwise. This means, `CreateHiveTableAsSelectCommand` from SQL CTAS should only trigger table insertion if table not exists.

For `DataFrameWriter.saveAsTable`, it can create the table if it doesn't exist, or append data to the table if `SaveMode` is `Append`. `DataFrameWriter.saveAsTable` will drop the table first if `SaveMode` is `Overwrite`. This means, `CreateHiveTableAsSelectCommand` from `DataFrameWriter.saveAsTable` either appends data to an existing table, or creates a new table.

In summary, the table insertion command triggered by `CreateHiveTableAsSelectCommand` should always set `overwrite` to false.

### Why are the changes needed?

fix a mistake

### Does this PR introduce _any_ user-facing change?

No. If the table does not exist, the `overwrite` flag doesn't matter. But it could be a potential bug in the future.

### How was this patch tested?

existing tests

Closes #39365 from cloud-fan/hive.

Lead-authored-by: Wenchen Fan <wenchen@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>