[SPARK-41861][SQL] Make v2 ScanBuilders' build() return typed scan
### What changes were proposed in this pull request?
The `ScanBuilder` interface has the `build()` method to return `Scan` objects. All the different implementations of it return scans that are of the type of the builder itself. Eg `ParquetScanBuilder` will return a `ParquetScan`, `TextScanBuilder` a `TextScan` etc. However, in the method overrides declaration, the return type is not changed to the more specific type but left as a generic `Scan`. For ease of development I'm proposing that the the `build()` methods of each implementation of `ScanBuilder`s returned a more strictly typed object instead of the generic `Scan` object

### Why are the changes needed?
When developing using spark as a library, it can be problematic to deal with the scan types as often manual casting is required when we should know for a fact that typed `ScanBuilder`s return typed scans. In order to avoid having to hard cast objects and have stricter requirements on the return types (We want a `ParquetScanBuilder` to return `ParquetScan`s only) these changes would be really helpful.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
No need to test as we are just changing the declared return type to the effective return type of the scanbuilders.

This change is safe because we are making the return type more precise and they are just reflecting what the methods return anyways

Closes #39367 from LorenzoMartini/lmartini/scan-types.

Authored-by: lmartini <lmartini@palantir.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>