[SPARK-41937][R] Fix error in R (>= 4.2.0) for SparkR datetime column comparing with Sys.time()
### What changes were proposed in this pull request?
1. Added the base R `all` function (or `any`, as appropriate) while doing the check of whether `class(.)` is `Column` or not; e.g., `if(all(class(.) == "Column"))`, or `if(inherits(., "Column"))` is proposed instead of the current approach `if(class(.) == "Column")`.
2. Also added relevant SparkR test cases to verify no warning or error is resulted on the updated functions.
3. Only R scripts have been modified.

### Why are the changes needed?
Base R 4.2.0 introduced a change ([[Rd] R 4.2.0 is released](https://stat.ethz.ch/pipermail/r-announce/2022/000683.html)), "Calling if() or while() with a condition of length greater than one gives an error rather than a warning."

The below code is a reproducible example of the issue. If it is executed in R >=4.2.0 then it will generate an error, or else just a warning message. `Sys.time()` is a multi-class object in R, and throughout the Spark R repository 'if' statement is used as: `if(class == "Column")` - this causes error in the latest R version >= 4.2.0. Note that R allows an object to have multiple 'class' names as a character vector ([R: Object Classes](https://stat.ethz.ch/R-manual/R-devel/library/base/html/class.html)); hence this type of check itself was not a good idea in the first place.

{
 SparkR::sparkR.session()
 t <- Sys.time()
 sdf <- SparkR::createDataFrame(data.frame(x = t + c(-1, 1, -1, 1, -1)))
 SparkR::collect(SparkR::filter(sdf, SparkR::column('x') > t))
}
#> Warning in if (class(e2) == 'Column') {: the condition has length > 1
#> and only the first element will be used
#> x
#> 1 2023-01-07 20:40:20
#> 2 2023-01-07 20:40:20

{
 Sys.setenv(`_R_CHECK_LENGTH_1_CONDITION_` = "true")
 SparkR::sparkR.session()
 t <- Sys.time()
 sdf <- SparkR::createDataFrame(data.frame(x = t + c(-1, 1, -1, 1, -1)))
 SparkR::collect(SparkR::filter(sdf, SparkR::column('x') > t))
}
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x'
#> in selecting a method for function 'collect': error in evaluating the
#> argument 'condition' in selecting a method for function 'filter': the
#> condition has length > 1

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Added multiple relevant SparkR unit test cases to verify the fix; the Jira ticket number is added in the new test_that block.

Closes #39454 from atalv/r42-comparetime-fixclass.

Authored-by: Vivek Atal <atalvivek@yahoo.co.in>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>