[SPARK-41877][CONNECT][TESTS] Reenable unpivot test and separate negative test
### What changes were proposed in this pull request?

This PR proposes to reenable unpivot test cases, and separate negative test cases that fail with a different result. I filed a separate JIRA for that.

Additionally, I also added one assert to be same as the regular PySpark side.

### Why are the changes needed?

To make sure on the test coverage, and feature parity.

### Does this PR introduce _any_ user-facing change?

No, test-only. Adds an assert to be consistent with the regular PySpark side.

### How was this patch tested?

Unittests were factored out. CI in this PR should test this out.

Closes #39482 from HyukjinKwon/SPARK-41877.

Lead-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>