[SPARK-41432][UI][FOLLOWUP] Fix a bug in protobuf serializer of SparkPlanGraphNodeWrapper
### What changes were proposed in this pull request?

SparkPlanGraphNodeWrapper can only contain either a node or a cluster. In the current implementation, both the node and cluster fields are not null. It breaks the assertion of the method `toSparkPlanGraphNode`:
```scala
class SparkPlanGraphNodeWrapper(
    val node: SparkPlanGraphNode,
    val cluster: SparkPlanGraphClusterWrapper) {

  def toSparkPlanGraphNode(): SparkPlanGraphNode = {
    assert(node == null ^ cluster == null, "Exactly one of node, cluster values to be set.")
    if (node != null) node else cluster.toSparkPlanGraphCluster()
  }

}
```
This PR is to fix the bug by using `oneof` in the protobuf definition of `SparkPlanGraphNodeWrapper`
### Why are the changes needed?

Bug fix

### Does this PR introduce _any_ user-facing change?

No
### How was this patch tested?

Update related UT. Also run SQLMetricsSuite with RocksDB backend enabled.

Closes #39471 from gengliangwang/fixGraph.

Authored-by: Gengliang Wang <gengliang@apache.org>
Signed-off-by: Gengliang Wang <gengliang@apache.org>