[SPARK-42037][INFRA] Rename `AMPLAB_` to `SPARK_` in Jenkins build environment variables
### What changes were proposed in this pull request?

This PR aims to rename the following environment variables and some legacy variable `amplap` from Apache Spark 3.4.0.
- `AMPLAB_JENKINS` -> `SPARK_JENKINS`
- `AMPLAB_JENKINS_BUILD_TOOL` -> `SPARK_JENKINS_BUILD_TOOL`
- `AMPLAB_JENKINS_BUILD_SCALA_PROFILE` -> `SPARK_JENKINS_BUILD_SCALA_PROFILE`
- `AMPLAB_JENKINS_BUILD_PROFILE` -> `SPARK_JENKINS_BUILD_PROFILE`

### Why are the changes needed?

As of today, Apache Spark's the only Jenkins farm is `Apple Silicon Jenkins` Farm on `Scaleway`.
- https://spark.apache.org/developer-tools.html

It's weird to use `AMPLAB_XXX` in new irrelevant environment.

### Does this PR introduce _any_ user-facing change?

No, this is a dev-only change for new branch-3.4 and master.

### How was this patch tested?

Pass the CIs.

Actually, there is no occurrence of `AMPLAB_` after this PR.
```
$ git grep AMPLAB_ | wc -l
       0
```

Closes #39539 from dongjoon-hyun/SPARK-42037.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>