[SPARK-42039][SQL] SPJ: Remove Option in KeyGroupedPartitioning#partitionValuesOpt
### What changes were proposed in this pull request?

Currently `KeyGroupedPartitioning#partitionValuesOpt` is of type: `Option[Seq[InternalRow]]`. This refactors it into
`Seq[InternalRow]`.

### Why are the changes needed?

It's unnecessary to use `Option` for the field. Originally I was thinking to use `None` for the case when all the input partitions are implicitly matched, so that we can skip comparing them in `EnsureRequirements`. However, I think it is not really a use case, and we can instead use `Seq.empty` for that if it comes up.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests.

Closes #39540 from sunchao/SPARK-42039.

Authored-by: Chao Sun <sunchao@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>