[SPARK-18011] Fix SparkR NA date serialization
### What changes were proposed in this pull request?
This PR ensures that SparkR serializes `NA` dates as `"NA"` (string) to avoid an undefined length when deserializing in the JVM.

### Why are the changes needed?
Currently, SparkR assumes that a `NegativeArraySizeException` when deserializing dates and timestamps represents an `NA`.  However, this handling can be made more robust by ensuring that serialization on the R side always provides a valid string length (note that `nchar(as.character(as.Date(NA)))` is `NA`).

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?

Closes #39681 from joveyuan-db/SPARK-18011.

Authored-by: Jove Yuan <jove.yuan@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>