[SPARK-42167][INFRA] Improve GitHub Action `lint` job to stop on failures earlier
### What changes were proposed in this pull request?

This PR aims to improve `GitHub Action` **lint** job to stop on failure earlier.

### Why are the changes needed?

Currently, we designed to put all installation steps before testing. However, we don't need to install `SparkR` dependency if the PR fails at `LICENSE` check.

By pairing the installation and testing steps, we can skip as much as possible in case of failures.

### Does this PR introduce _any_ user-facing change?

No. This is a INFRA change.

### How was this patch tested?

Manual review.

Closes #39716 from dongjoon-hyun/SPARK-42167.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>