[SPARK-41775][PYTHON][FOLLOWUP] Fix stdout rerouting
### What changes were proposed in this pull request?

This is a follow-up of https://github.com/apache/spark/pull/39369 which aims to fix `stderr` rerouting to `stdout` which is required for users seeing any errors that come up due to distributed training,

### Why are the changes needed?

Previously, the error would be lost since it's never returned to the user. We are fixing that issue.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

N/A

Closes #39724 from rithwik-db/sterr-followup.

Authored-by: Rithwik Ediga Lakhamsani <rithwik.ediga@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>